backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "road-brain-backend"
version = "0.1.0"
description = "FastAPI backend with Supabase auth"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "supabase>=2.6.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.1",
    "email-validator>=2.2.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]
